+++ pytraj/run_tests.py
@@ -3,8 +3,10 @@ from pytraj.base import *
 from pytraj import io as mdio
 from pytraj.utils.check_and_assert import assert_almost_equal
 from .data_sample.load_sample_data import load_sample_data
+from pytraj import adict
+from pytraj.misc import action_help
 
-def run_tests(self):
+def run_tests():
     traj = load_sample_data()
 
     # test Topology
@@ -13,6 +15,8 @@ def run_tests(self):
     print (traj)
     # test import action
     action_help()
+    print ("all cpptraj actions")
+    print (adict.keys())
 
     print ("OK")